SubTasks done

+++ app/src/main/java/com/kunano/tasks_to_do/core/data/model/entities/LocalCategoryEntity.kt
@@ -6,6 +6,6 @@ import androidx.room.PrimaryKey
 
 @Entity
 class LocalCategoryEntity(
-    @PrimaryKey(autoGenerate = true) val categoryId: Int? = null,
+    @PrimaryKey(autoGenerate = true) val categoryId: Long? = null,
     var categoryName: String
 )
